+++ clients/csal/test/main.c
@@ -3,231 +3,7 @@
 #include<string.h>
 #include "csal/csal.h"
 
-#define TC_TEST( tf, expr ) test_case_test( tf, expr, #expr, __FILE__, __LINE__ )
-
-#define TC_TEST_EQUAL_INT( tc, val, expected ) test_case_equal_int( tc, val, expected, __FILE__, __LINE__ )
-#define TC_TEST_EQUAL_FLOAT( tc, val, expected ) test_case_equal_float( tc, val, expected, __FILE__, __LINE__ )
-#define TC_TEST_EQUAL_DOUBLE( tc, val, expected ) test_case_equal_double( tc, val, expected, __FILE__, __LINE__ )
-
-#define TS_TEST_CASE( tf, f ) test_suite_add_test_case( tf, #f, f )
-
-struct _test_suite_t;
-struct _test_case_t;
-
-typedef int (*TEST_CASE_FUNCTION)( struct _test_case_t* );
-
-enum TEST_SUITE_CTRL_FLAGS
-{
-    TEST_SUITE_CTRL_NONE=0
-        , TEST_SUITE_CTRL_EXIT_ON_FAIL=1
-        , TEST_SUITE_CTRL_VERBOSE=2
-};
-
-
-typedef struct _test_case_t
-{
-    TEST_CASE_FUNCTION f;
-    char szname[ 256];
-    int npassed;
-    int nfailed;
-
-    int ctrl_flags;
-    struct _test_suite_t* test_suite;
-
-} test_case_t;
-
-typedef struct _test_suite_t
-{
-    unsigned int ctrl_flags;
-    int ntestcases;
-    test_case_t* testcases;
-
-} test_suite_t;
-
-
-
-int test_case_init( test_case_t* self, test_suite_t* owner )
-{
-    int err = 0;
-    
-    memset( self, 0, sizeof( test_case_t ) );
-
-    self->test_suite = owner;
-
-    return err;
-}
-
-int test_case_name_set( test_case_t* self, const char* pszname )
-{
-    int err =0;
-
-    strncpy( self->szname, pszname, 255 );
-
-    return err;
-}
-
-char* test_case_name_get( test_case_t* self )
-{
-    return self->szname;
-}
-
-int test_case_finish( test_case_t* self )
-{
-    int err = 0;
-
-    return err;
-}
-
-int test_suite_add_test_case( test_suite_t* self, const char* pszname, TEST_CASE_FUNCTION f)
-{
-    int err = 0;
-
-    test_case_t*  tc = NULL;
-
-
-    self->ntestcases++;
-    self->testcases = (test_case_t*)realloc( self->testcases, sizeof(test_case_t) * self->ntestcases );
-
-    tc = self->testcases + self->ntestcases-1;
-
-    test_case_init( tc, self );
-    test_case_name_set( tc, pszname );
-    tc->f = f;
-
-
-    printf( "\nTest case : %s : START\n", pszname );
-
-    tc->f( tc );
-    
-    printf( "Test Case : %s : %d/%d passed\n",pszname,  tc->npassed, tc->npassed + tc->nfailed );
-    printf( "Test Case : %s : %d/%d failed\n", pszname, tc->nfailed, tc->npassed + tc->nfailed );
-    printf( "Test case : %s : END\n", pszname );
-
-    return err;
-}
-
-int test_suite_init( test_suite_t* self, unsigned int ctrl_flags )
-{
-    self->ctrl_flags = ctrl_flags;
-
-    self->ntestcases = 0;
-    self->testcases = NULL;
-    return 0;
-}
-
-static int _test_case_passed( test_case_t* self, int res, const char* expr, const char* pszfile, unsigned int lineno )
-{
-    int err = 0;
-
-    self->npassed++;
-
-    printf( "%s(%d) : %s : %s\n", pszfile, lineno, expr, "PASS" );
-
-    return err;
-}
-
-static int _test_case_failed( test_case_t* self, int res, const char* expr, const char* pszfile, unsigned int lineno )
-{
-    int err = 0;
-
-    self->nfailed++;
-
-    printf( "%s(%d) : %s : %s\n", pszfile, lineno, expr, "FAIL" );
-
-    if( !res && self->ctrl_flags & TEST_SUITE_CTRL_EXIT_ON_FAIL )
-        exit(1);
-
-    return err;
-}
-
-
-
-int test_case_test( test_case_t* self, int res, const char* expr, const char* file, unsigned int lineno )
-{
-    if( res )
-        _test_case_passed( self, res, expr, file, lineno );
-    else
-        _test_case_failed( self, res, expr, file, lineno );
-
-    return 0;
-}
-
-int test_case_equal_int( test_case_t* self, int val, int expected, const char* pszfile, unsigned int lineno )
-{
-    int res = (val == expected);
-
-    char buffer[ 512];
-
-    sprintf( buffer, "%d == %d", val, expected );
-
-    if( res )
-        _test_case_passed( self, res, buffer, pszfile, lineno );
-    else
-        _test_case_failed( self, res, buffer, pszfile, lineno );
-
-    return 0;
-}
-
-int test_case_equal_float( test_case_t* self, float val, float expected, const char* pszfile, unsigned int lineno )
-{
-    int res = (val == expected);
-
-    char buffer[ 512];
-
-    sprintf( buffer, "%f == %f", val, expected );
-
-    if( res )
-        _test_case_passed( self, res, buffer, pszfile, lineno );
-    else
-        _test_case_failed( self, res, buffer, pszfile, lineno );
-
-    return 0;
-}
-
-int test_case_equal_double( test_case_t* self, double val, double expected, const char* pszfile, unsigned int lineno )
-{
-    int res = (val == expected);
-
-    char buffer[ 512];
-
-    sprintf( buffer, "%f == %f", val, expected );
-
-    if( res )
-        _test_case_passed( self, res, buffer, pszfile, lineno );
-    else
-        _test_case_failed( self, res, buffer, pszfile, lineno );
-
-    return 0;
-}
-
-
-int test_suite_finish( test_suite_t* self )
-{
-    size_t i = 0;
-
-
-    int npassed=0;
-    int nfailed=0;
-
-    for( i = 0; i < self->ntestcases; ++i )
-    {
-        test_case_t* tc = self->testcases + i;
-    
-        npassed += tc->npassed;
-        nfailed += tc->nfailed;
-
-        test_case_finish( tc );
-    }
-
-    printf( "\n" );
-    printf( "%d/%d tests failed\n", nfailed, nfailed+npassed );
-    printf( "%d/%d tests passed\n", npassed, nfailed+npassed );
-    printf( "%d test cases\n", self->ntestcases );
-
-    free( self->testcases );
-
-    return 0;
-}
+#include "test_suite.h"
 
 
 int float_cmp( float val1, float val2, float tol )
